refactor : 상위 디렉토리 이름 content -> post
통일성은 항상 중요합니다
