+++ src/main/java/com/sparksInTheStep/webBoard/post/domain/TimeStamp.java
@@ -1,9 +1,8 @@
-package com.sparksInTheStep.webBoard.content.domain;
+package com.sparksInTheStep.webBoard.post.domain;
 
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
